requirements.txt
click>=8.0.0
requests>=2.25.0
colorama>=0.4.4
openai>=1.0.0
pyyaml>=6.0